Remove build query and rename user package to politician
