+++ src/politician/extractor.py
